# streamhub/shared/openapi.yaml
openapi: 3.1.0
info:
  title: StreamHub API
  version: 0.1.0
  description: >
    The contract every StreamHub client (web, Roku, Android TV) is built against.
    Phase 1 covers auth, the video library, playback decisions, progress, and
    HLS streaming. Music/photos (Phase 3) and cloud/ABR (Phase 6) extend this.
servers:
  - url: http://localhost:8080
    description: Local development

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
    User:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        isAdmin: { type: boolean }
        createdAt: { type: integer, format: int64 }
    LoginResponse:
      type: object
      properties:
        token: { type: string }
        user: { $ref: '#/components/schemas/User' }
    LibraryItem:
      type: object
      properties:
        id: { type: string }
        type:
          type: string
          enum: [movie, episode, track, album, photo, home_video]
        parentId: { type: string }
        title: { type: string }
        sortTitle: { type: string }
        year: { type: integer }
        overview: { type: string }
        container: { type: string }
        videoCodec: { type: string }
        audioCodec: { type: string }
        width: { type: integer }
        height: { type: integer }
        durationSeconds: { type: number }
        sizeBytes: { type: integer, format: int64 }
        posterUrl: { type: string }
        addedAt: { type: integer, format: int64 }
        updatedAt: { type: integer, format: int64 }
    PlaybackDecision:
      type: object
      properties:
        directPlay: { type: boolean }
        reason: { type: string }
        maxHeight: { type: integer }
    Progress:
      type: object
      properties:
        itemId: { type: string }
        positionSeconds: { type: number }
        updatedAt: { type: integer, format: int64 }

security:
  - bearerAuth: []

paths:
  /healthz:
    get:
      summary: Liveness probe
      security: []
      responses:
        '200':
          description: Service is up
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }

  /api/auth/login:
    post:
      summary: Exchange credentials for an access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /api/me:
    get:
      summary: Current authenticated user
      responses:
        '200':
          description: The user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }

  /api/items:
    get:
      summary: List library items
      parameters:
        - in: query
          name: type
          schema: { type: string }
          description: Filter by item type
        - in: query
          name: parent
          schema: { type: string }
          description: Filter by parent (collection) ID
      responses:
        '200':
          description: Matching items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/LibraryItem' }

  /api/items/{id}:
    get:
      summary: Get a single item
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: The item
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LibraryItem' }
        '404':
          description: Not found

  /api/items/{id}/playback-info:
    get:
      summary: Get the direct-play/transcode decision for an item
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Decision
          content:
            application/json:
              schema:
                type: object
                properties:
                  itemId: { type: string }
                  decision: { $ref: '#/components/schemas/PlaybackDecision' }

  /api/items/{id}/progress:
    get:
      summary: Get playback progress for the current user
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Progress (positionSeconds is 0 if none recorded)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Progress' }
    put:
      summary: Save playback progress for the current user
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [positionSeconds]
              properties:
                positionSeconds: { type: number }
      responses:
        '204':
          description: Saved

  /api/items/{id}/hls.m3u8:
    get:
      summary: Start (or reuse) an HLS session and redirect to its playlist
      description: >
        Starts an FFmpeg HLS session for the item (direct-play remux or
        transcode, per the playback decision) and 302-redirects to the session
        playlist. The token may be supplied via the `token` query parameter for
        players that cannot set an Authorization header.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: token
          schema: { type: string }
      responses:
        '302':
          description: Redirect to the session media playlist
        '404':
          description: Item not found
        '504':
          description: Stream failed to start in time

  /api/stream/sessions/{sid}/{file}:
    get:
      summary: Fetch an HLS playlist or segment for a session
      description: >
        Served without JWT — the unguessable session ID is the capability.
        Signed, time-limited URLs are planned for Phase 6.
      security: []
      parameters:
        - in: path
          name: sid
          required: true
          schema: { type: string }
        - in: path
          name: file
          required: true
          schema: { type: string }
      responses:
        '200':
          description: The playlist (application/vnd.apple.mpegurl) or a segment (video/mp2t)
        '404':
          description: Unknown session or file

  /api/library/scan:
    post:
      summary: Trigger a library scan (admin only)
      responses:
        '202':
          description: Scan started
        '403':
          description: Admin required